requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
Click==7.0
Cython==0.29.8
envparse==0.2.0
idna==2.8
libinjection-python==1.1.4
msgpack==0.6.1
multidict==4.5.2
nanoid==2.0.0
pycrypto==2.6.1
PyYAML==5.1
systemd==0.16.1
yarl==1.3.0
